+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+
+- Prometheus metrics for trial garbage collection, trial duration and tick duration
+
 ## v1.0.2 - 2021-07-07
 
 ### Changed